Customer-Churn/requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
mysql-connector-python
pymssql
kagglehub[hf-datasets]
kaggle
google-cloud-storage
schedule
reportlab
xgboost
feast
psycopg2
psycopg
snowflake-connector-python
google-cloud-secret-manager
apache-airflow
-e.